Move loggers to MagicOnion.Server.Diagnostics
