+++ src/MagicOnion.Server/Hubs/Group.ConcurrentDictionary.cs
@@ -1,3 +1,4 @@
+using MagicOnion.Server.Diagnostics;
 using MagicOnion.Utils;
 using MessagePack;
 using System.Collections.Concurrent;
