+++ src/MagicOnion.Server/Hubs/Group.cs
@@ -1,5 +1,6 @@
 using System.Collections.Concurrent;
 using System.Diagnostics.CodeAnalysis;
+using MagicOnion.Server.Diagnostics;
 using MessagePack;
 
 namespace MagicOnion.Server.Hubs;
